tests: extract tests to multiple files
